% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{prepare the matrix and cell labels for training}
\usage{
getMatrix(obj, column, state1, state0, prefix = "./train", species = "human")
}
\arguments{
\item{obj}{SeuratObject}

\item{column}{which column of cell labels}

\item{state1}{target cell type}

\item{state0}{non-target cell state}

\item{prefix}{prefix of file to save data}

\item{species}{human or mouse}
}
\value{
save matrix and cell information as prefix_data.csv and prefix_info.csv
}
\description{
prepare the matrix and cell labels for training
}
\examples{
\dontrun{
library(Seurat)
getMatrix(pbmc_small,column = 'groups', state1 = "g2", state0 = "g1")
}
}
