% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTopWeights.R
\name{printTopWeights}
\alias{printTopWeights}
\title{print top features for specific state}
\usage{
printTopWeights(
  stateName = "EMT",
  w1_file = NULL,
  myStateName = "newState",
  species = "human",
  showN = 50,
  print = T
)
}
\arguments{
\item{stateName}{cell state to print}

\item{w1_file}{if set, stateName will be ignored}

\item{myStateName}{state name of w1_file}

\item{species}{human or mouse for now}

\item{showN}{show n features}
}
\value{
a list of positive and bottom features
}
\description{
print top features for specific state
}
\examples{
\dontrun{
topFeaturesList <- printTopWeights(stateName = "EMT")
}
}
