% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCellStates.R
\name{scoreStates}
\alias{scoreStates}
\title{the R implemention of a computational framework to evaluate the cellular states}
\usage{
scoreStates(
  mat,
  state = c("Exhaustion", "EMT", "CellCycle", "Hypoxia"),
  species = "human",
  n.chunks = min(10, ncol(mat))
)
}
\arguments{
\item{mat}{a features x barcodes expression matrix}

\item{state}{cell state(s) to evaluate}

\item{species}{human or mouse for now}

\item{n.chunks}{split these cells into n.chunks batch}
}
\value{
a scores matrix
}
\description{
score the cell states on the single-cell RNA-seq digital expression profiles
}
\examples{
\dontrun{
library(Seurat)
myscores <- scoreStates(pbmc_small@assays$RNA@counts, state = "CellCycle")
pbmc_small = AddMetaData(pbmc_small, metadata = myscores)
}
}
\references{
Quantitative Learning of Cellular Features From Single-cell Transcriptomics Data Facilitates Effective Drug Repurposing
}
