% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreStates_selftrain.R
\name{scoreStates_selftrain}
\alias{scoreStates_selftrain}
\title{the R implemention of a computational framework to evaluate the cellular states}
\usage{
scoreStates_selftrain(
  mat,
  w1_file,
  b1_file,
  add_state_name = "mystate",
  species = "human",
  n.chunks = min(10, ncol(mat))
)
}
\arguments{
\item{mat}{a features x barcodes expression matrix}

\item{w1_file}{weight file downloaded from the online training website}

\item{b1_file}{bias file downloaded from the online training website}

\item{add_state_name}{cell state name}

\item{species}{support human or mouse for now}

\item{n.chunks}{split these cells into n.chunks batch}
}
\value{
a scores matrix
}
\description{
score the cell states based on your model
}
\examples{
\dontrun{
library(Seurat)
myscores <- scoreStates_selftrain(pbmc_small@assays$RNA@counts,
    w1_file = system.file('extdata','w1.csv',package = 'rSuperFeat'),
    b1_file = system.file('extdata','b1.csv',package = 'rSuperFeat'),
    species = "human")
pbmc_small = AddMetaData(pbmc_small, metadata = myscores)
# or on your model
myscores <- scoreStates_selftrain(pbmc_small@assays$RNA@counts,
    w1_file = "w1.csv", b1_file = "b1.csv", species = "human")
pbmc_small = AddMetaData(pbmc_small, metadata = myscores)

}
}
\references{
Quantitative Learning of Cellular Features From Single-cell Transcriptomics Data Facilitates Effective Drug Repurposing
}
