% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCellStates.R
\name{scoreStates}
\alias{scoreStates}
\title{score the cell states in cells in tumor micro-environment based on the single-cell RNA-seq digital expression profiles}
\usage{
scoreStates(
  mat,
  state = c("Exhaustion", "EMT", "CellCycle", "Hypoxia", "Cytotoxic"),
  normalize = TRUE,
  n.chunks = min(100, ncol(mat))
)
}
\arguments{
\item{mat}{a features x barcodes expression matrix}

\item{state}{cell state(s) to evaluate}

\item{normalize}{whether to normalize the matrix to 0-1 range}

\item{n.chunks}{split these cells into n.chunks batch}
}
\value{
a scores matrix
}
\description{
score the cell states in cells in tumor micro-environment based on the single-cell RNA-seq digital expression profiles
}
\examples{
\dontrun{
myscores <- scoreStates(pbmc_small@assays$RNA@data)
pbmc_small = AddMetaData(pbmc_small, metadata = myscores)
}
}
\references{
SuperFeat: A Framework of Quantitative Feature Learning and Assessment from Single-cell Transcriptomics Data
}
