% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCellStates_v2.R
\name{scoreStates_selftrain}
\alias{scoreStates_selftrain}
\title{score the cell states in cells in tumor micro-environment based on the single-cell RNA-seq digital expression profiles}
\usage{
scoreStates_selftrain(
  mat,
  w1_file = "extdata/w1.csv",
  b1_file = "extdata/b1.csv",
  cell_state_name = NULL,
  normalize = TRUE,
  n.chunks = min(100, ncol(mat))
)
}
\arguments{
\item{mat}{a features x barcodes expression matrix}

\item{w1_file}{weight file downloaded from the online training website}

\item{b1_file}{bias file downloaded from the online training website}

\item{cell_state_name}{cell state name}

\item{normalize}{whether to normalize the matrix to 0-1 range}

\item{n.chunks}{split these cells into n.chunks batch}
}
\value{
a scores matrix
}
\description{
score the cell states in cells in tumor micro-environment based on the single-cell RNA-seq digital expression profiles
}
\examples{
\dontrun{
myscores <- scoreStates_selftrain(pbmc_small@assays$RNA@data,
    w1_file = system.file('extdata','w1.csv',package = 'rSuperFeat'),
    b1_file = system.file('extdata','b1.csv',package = 'rSuperFeat'))
}
}
\references{
SuperFeat: A Framework of Quantitative Feature Learning and Assessment from Single-cell Transcriptomics Data
}
